Put new functions in new file
